% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DESeq2Test}
\alias{DESeq2Test}
\title{Differential expression using DESeq2}
\usage{
DESeq2Test(sub_data, min_gene_expressed, min_valid_cells,
  contrast = unique(sub_data$compare_group), test = "Wald",
  fitType = "parametric", sfType = "ratio", betaPrior = FALSE,
  quiet = FALSE, modelMatrixType = "standard",
  minReplicatesForReplace = 7, useT = FALSE, minmu = 0.5,
  parallel = FALSE, BPPARAM = bpparam())
}
\arguments{
\item{sub_data}{Count data removed cell_type and selected certain two
compare_group}

\item{min_gene_expressed}{Genes expressed in minimum number of cells}

\item{min_valid_cells}{Minimum number of genes detected in the cell}

\item{contrast}{String vector specifying the contrast to be
tested against the log2-fold-change threshold}

\item{test}{either "Wald" or "LRT", which will then use either
Wald significance tests (defined by \code{\link{nbinomWaldTest}}),
or the likelihood ratio test on the difference in deviance between a
full and reduced model formula (defined by \code{\link{nbinomLRT}})}

\item{fitType}{either "parametric", "local", or "mean"
for the type of fitting of dispersions to the mean intensity.
See \code{\link{estimateDispersions}} for description.}

\item{sfType}{either "ratio", "poscounts", or "iterate"
for teh type of size factor estimation. See
\code{\link{estimateSizeFactors}} for description.}

\item{betaPrior}{whether or not to put a zero-mean normal prior on
the non-intercept coefficients
See \code{\link{nbinomWaldTest}} for description of the calculation
of the beta prior. In versions \code{>=1.16}, the default is set
to \code{FALSE}, and shrunken LFCs are obtained afterwards using
\code{\link{lfcShrink}}.}

\item{quiet}{whether to print messages at each step}

\item{modelMatrixType}{either "standard" or "expanded", which describe
how the model matrix, X of the GLM formula is formed.
"standard" is as created by \code{model.matrix} using the
design formula. "expanded" includes an indicator variable for each
level of factors in addition to an intercept. for more information
see the Description of \code{\link{nbinomWaldTest}}.
betaPrior must be set to TRUE in order for expanded model matrices
to be fit.}

\item{minReplicatesForReplace}{the minimum number of replicates required
in order to use \code{\link{replaceOutliers}} on a
sample. If there are samples with so many replicates, the model will
be refit after these replacing outliers, flagged by Cook's distance.
Set to \code{Inf} in order to never replace outliers.}

\item{useT}{logical, passed to \code{\link{nbinomWaldTest}}, default is FALSE,
where Wald statistics are assumed to follow a standard Normal}

\item{minmu}{lower bound on the estimated count for fitting gene-wise dispersion
and for use with \code{nbinomWaldTest} and \code{nbinomLRT}}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel
execution using \code{BiocParallel}, see next argument \code{BPPARAM}.
A note on running in parallel using \code{BiocParallel}: it may be
advantageous to remove large, unneeded objects from your current
R environment before calling \code{DESeq},
as it is possible that R's internal garbage collection
will copy these files while running on worker nodes.}

\item{BPPARAM}{an optional parameter object passed internally
to \code{\link{bplapply}} when \code{parallel=TRUE}.
If not specified, the parameters last registered with
\code{\link{register}} will be used.}
}
\value{
A matrix of differentially expressed genes and related statistics.
}
\description{
Identifies differentially expressed genes between two groups of cells using
DESeq2
}
\details{
This test does not support pre-processed genes. To use this method, please
install DESeq2, using the instructions at
https://bioconductor.org/packages/release/bioc/html/DESeq2.html
}
\references{
Love MI, Huber W and Anders S (2014). "Moderated estimation of
fold change and dispersion for RNA-seq data with DESeq2." Genome Biology.
https://bioconductor.org/packages/release/bioc/html/DESeq2.html
}
