% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DESingleTest}
\alias{DESingleTest}
\title{Differential expression using DEsingle}
\usage{
DESingleTest(sub_data, min_gene_expressed, min_valid_cells,
  contrast = unique(sub_data$compare_group), parallel = FALSE,
  BPPARAM = bpparam())
}
\arguments{
\item{sub_data}{Count data removed cell_type and selected certain two
compare_group}

\item{min_gene_expressed}{Genes expressed in minimum number of cells}

\item{min_valid_cells}{Minimum number of genes detected in the cell}

\item{contrast}{String vector specifying the contrast to be
tested against the log2-fold-change threshold}

\item{parallel}{If FALSE (default), no parallel computation is used;
if TRUE, parallel computation using \code{BiocParallel}, with argument
\code{BPPARAM}.}

\item{BPPARAM}{An optional parameter object passed internally to
\code{\link{bplapply}} when \code{parallel=TRUE}. If not specified,
\code{\link{bpparam}()} (default) will be used.}
}
\value{
A matrix of differentially expressed genes and related statistics.
}
\description{
Identifies differentially expressed genes between two groups of cells using
DEsingle
}
\details{
This test does not support pre-processed genes. To use this method, please
install DEsingle, using the instructions at
https://github.com/miaozhun/DEsingle
}
\references{
Zhun Miao, Ke Deng, Xiaowo Wang, Xuegong Zhang (2018). DEsingle
for detecting three types of differential expression in single-cell RNA-seq
data. Bioinformatics, bty332. 10.1093/bioinformatics/bty332.
}
