% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{edgeRTest}
\alias{edgeRTest}
\title{Differential expression using edgeR}
\usage{
edgeRTest(sub_data, min_gene_expressed, min_valid_cells,
  contrast = unique(sub_data$compare_group), calcNormMethod = "TMM",
  trend.method = "locfit", tagwise = TRUE, robust = FALSE)
}
\arguments{
\item{sub_data}{Count data removed cell_type and selected certain two
compare_group}

\item{min_gene_expressed}{Genes expressed in minimum number of cells}

\item{min_valid_cells}{Minimum number of genes detected in the cell}

\item{contrast}{String vector specifying the contrast to be
tested against the log2-fold-change threshold}

\item{calcNormMethod}{normalization method to be used}

\item{trend.method}{method for estimating dispersion trend. Possible values
are "none", "movingave", "loess" and "locfit" (default).}

\item{tagwise}{logical, should the tagwise dispersions be estimated}

\item{robust}{logical, should the estimation of prior.df be robustified
against outliers}
}
\value{
A matrix of differentially expressed genes and related statistics.
}
\description{
Identifies differentially expressed genes between two groups of cells using
edgeR
}
\details{
This test does not support pre-processed genes. To use this method, please
install edgeR, using the instructions at
http://bioconductor.org/packages/release/bioc/html/edgeR.html
}
\references{
McCarthy, J. D, Chen, Yunshun, Smyth, K. G (2012). “Differential
expression analysis of multifactor RNA-Seq experiments with respect to
biological variation.” Nucleic Acids Research, 40(10), 4288-4297.

Robinson MD, McCarthy DJ, Smyth GK (2010). “edgeR: a Bioconductor
package for differential expression analysis of digital gene expression data.”
Bioinformatics, 26(1), 139-140.
https://github.com/cole-trapnell-lab/monocle-release
}
