% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindLR.R
\name{FindLR}
\alias{FindLR}
\title{Finding ligand-receptor pairs}
\usage{
FindLR(data_1, data_2 = NULL, datatype, comm_type, database = NULL)
}
\arguments{
\item{data_1}{Data used to find the ligand-receptor pairs}

\item{data_2}{Second dataset used to find ligand-receptor pairs. If set NULL,
paris will be found within data_1. Otherwise, pairs will be found between
data_1 and data_2. Default is NULL.}

\item{datatype}{Type of data used as input. Options are "mean count"
and "DEG"}

\item{comm_type}{Communication type. Available options are "cytokine",
"checkpoint", "growth factor", "other"}

\item{database}{Database used to find ligand-receptor pairs. If set NULL,
the build-in database will be used.}
}
\value{
A dataframe of the significant interactions
}
\description{
This function loads the highly expressed genes or differentail expressed
genes as a dataframe. Significant interactions are found through mapping
these genes to our ligand-receptor database.
}
\references{
Cytokines, Inflammation and Pain. Zhang et al,2007.

Cytokines, Chemokines and Their Receptors. Cameron et al, 2000-2013

Robust prediction of response to immune checkpoint blockade therapy
in metastatic melanoma. Auslander et al, 2018.

A draft network of ligand-receptor-mediated multicellular signalling
in human, Jordan A. Ramilowski, Nature Communications, 2015
}
