% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{MASTTest}
\alias{MASTTest}
\title{Differential expression using MAST}
\usage{
MASTTest(sub_data, min_gene_expressed, min_valid_cells,
  contrast = unique(sub_data$compare_group), method = "glm",
  silent = FALSE, check_logged = TRUE)
}
\arguments{
\item{sub_data}{Count data removed cell_type and selected certain two
compare_group}

\item{min_gene_expressed}{Genes expressed in minimum number of cells}

\item{min_valid_cells}{Minimum number of genes detected in the cell}

\item{contrast}{String vector specifying the contrast to be
tested against the log2-fold-change threshold}

\item{method}{Character vector, either ’glm’, ’glmer’ or ’bayesglm’}

\item{check_logged}{Set FALSE to override sanity checks that try to
ensure that the default assay is log-transformed and has at least one
exact zero}

\item{Silence}{Common problems with fitting some genes}
}
\value{
A matrix of differentially expressed genes and related statistics.
}
\description{
Identifies differentially expressed genes between two groups of cells using
MAST
}
\details{
To use this method, please install MAST, using the instructions at
https://github.com/RGLab/MAST
}
\references{
MAST: a flexible statistical framework for assessing transcriptional
changes and characterizing heterogeneity in single-cell RNA sequencing
data G Finak, A McDavid, M Yajima, J Deng, V Gersuk, AK Shalek, CK Slichter
et al Genome biology 16 (1), 278
}
