% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{MonocleTest}
\alias{MonocleTest}
\title{Differential expression using monocle}
\usage{
MonocleTest(sub_data, min_gene_expressed, min_valid_cells,
  contrast = unique(sub_data$compare_group), batch = NULL, cores = 4)
}
\arguments{
\item{sub_data}{Count data removed cell_type and selected certain two
compare_group}

\item{min_gene_expressed}{Genes expressed in minimum number of cells}

\item{min_valid_cells}{Minimum number of genes detected in the cell}

\item{contrast}{String vector specifying the contrast to be
tested against the log2-fold-change threshold}

\item{batch}{Different batch identifier}

\item{cores}{The number of cores to be used while testing each gene
for differential expression.}
}
\value{
A matrix of differentially expressed genes and related statistics.
}
\description{
Identifies differentially expressed genes between two groups of cells using
monocle
}
\details{
This test does not support pre-processed genes. To use this method, please
install monocle, using the instructions at
https://bioconductor.org/packages/release/bioc/html/monocle.html
}
\references{
Qiu X, Hill A, Packer J, Lin D, Ma Y, Trapnell C (2017).
“Single-cell mRNA quantification and differential analysis with Census.”
Nature Methods.
https://github.com/cole-trapnell-lab/monocle-release
}
