% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetView.R
\name{NetView}
\alias{NetView}
\title{Network Viewing of cell-cell communication}
\usage{
NetView(data, col, label = TRUE, edge.curved = 0.5, shape = "circle",
  layout = nicely(), vertex.size = 20, margin = 0.2,
  vertex.label.cex = 1.5, vertex.label.color = "black",
  arrow.width = 1.5, edge.label.color = "black", edge.label.cex = 1,
  edge.max.width = 10)
}
\arguments{
\item{data}{A dataframe containing ligand-receptor pairs and corresponding
cell typesused to do the plotting}

\item{col}{Colors used to represent different cell types}

\item{label}{Whether or not shows the label of edges (number of connections
between different cell types)}

\item{edge.curved}{Specifies whether to draw curved edges, or not.
This can be a logical or a numeric vector or scalar.
First the vector is replicated to have the same length as the number of
edges in the graph. Then it is interpreted for each edge separately.
A numeric value specifies the curvature of the edge; zero curvature means
straight edges, negative values means the edge bends clockwise, positive
values the opposite. TRUE means curvature 0.5, FALSE means curvature zero}

\item{shape}{The shape of the vertex, currently “circle”, “square”,
“csquare”, “rectangle”, “crectangle”, “vrectangle”, “pie” (see
vertex.shape.pie), ‘sphere’, and “none” are supported, and only by the
plot.igraph command. “none” does not draw the vertices at all, although
vertex label are plotted (if given). See shapes for details about vertex
shapes and vertex.shape.pie for using pie charts as vertices.}

\item{layout}{The layout specification. It must be a call to a layout
specification function.}

\item{vertex.size}{The size of vertex}

\item{margin}{The amount of empty space below, over, at the left and right
of the plot, it is a numeric vector of length four. Usually values between
0 and 0.5 are meaningful, but negative values are also possible, that will
make the plot zoom in to a part of the graph. If it is shorter than four
then it is recycled.}

\item{vertex.label.cex}{The label size of vertex}

\item{vertex.label.color}{The color of label for vertex}

\item{arrow.width}{The width of arrows}

\item{edge.label.color}{The color for single arrow}

\item{edge.label.cex}{The size of label for arrows}

\item{edge.max.width}{The maximum arrow size}
}
\value{
A network graph of the significant interactions
}
\description{
This function loads the significant interactions as a dataframe, and colors
represent different types of cells as a structure. The width of edges represent
the strength of the communication. Labels on the edges show exactly how many
interactions exist between two types of cells.
}
\references{
Csardi G, Nepusz T: The igraph software package for complex network
research, InterJournal, Complex Systems 1695. 2006.
http://igraph.org
}
