% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawParse.R
\name{rawParse}
\alias{rawParse}
\title{Parsing the data to get top expressed genes}
\usage{
rawParse(data, top_genes = 50, stats = "mean")
}
\arguments{
\item{data}{Input data, raw or normalized count with 'cell_type' column}

\item{top_genes}{(scale 1 to 100) Top percent highly expressed genes used
to find ligand-receptor pairs, default is 50}

\item{stats}{Whether calculates the mean or the median of the data. Available
options are 'mean' and 'median'.}
}
\value{
A dataframe of the data
}
\description{
This function loads the count data as a dataframe. It assumes that each line
contains gene expression profile of one single cell, and each column
contains the one single gene expression profile in different cells. The dataframe
should also contain the cell type information with column name 'cell_type'.
Group information should also be included as 'compare_group' if users want
to call differntial expressed ligand-receptor pairs. Batch information as
'batch' is optional. If included, users may want to use the raw count data
for later analysis.
}
