% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{SCDETest}
\alias{SCDETest}
\title{Differential expression using scde}
\usage{
SCDETest(sub_data, min_gene_expressed, min_valid_cells,
  contrast = unique(sub_data$compare_group), batch = NULL,
  n.randomizations = 150, n.cores = 10, batch.models = models,
  return.posteriors = FALSE, verbose = 1)
}
\arguments{
\item{sub_data}{Count data removed cell_type and selected certain two
compare_group}

\item{min_gene_expressed}{Genes expressed in minimum number of cells}

\item{min_valid_cells}{Minimum number of genes detected in the cell}

\item{contrast}{String vector specifying the contrast to be
tested against the log2-fold-change threshold}

\item{batch}{Different batch identifier}

\item{n.cores}{number of cores to utilize}

\item{batch.models}{(optional) separate models for the batch data (if generated
using batch-specific group argument). Normally the same models are used.}

\item{return.posteriors}{whether joint posterior matrices should be returned}

\item{verbose}{integer verbose level (1 for verbose)}

\item{@param}{n.randomizations number of bootstrap randomizations to be performed}
}
\value{
A matrix of differentially expressed genes and related statistics.
}
\description{
Identifies differentially expressed genes between two groups of cells using
scde
}
\details{
This test does not support pre-processed genes. To use this method, please
install scde, using the instructions at
http://hms-dbmi.github.io/scde/tutorials.html
}
\references{
"Bayesian approach to single-cell differential expression
analysis" (Kharchenko PV, Silberstein L, Scadden DT, Nature Methods,
doi:10.1038/nmeth.2967)
https://github.com/hms-dbmi/scde
}
