% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{WilcoxTest}
\alias{WilcoxTest}
\title{Differential expression using wilcox}
\usage{
WilcoxTest(sub_data, min_gene_expressed, min_valid_cells,
  contrast = unique(sub_data$compare_group), datatype = "raw count",
  verbose = 0)
}
\arguments{
\item{sub_data}{Count data removed cell_type and selected certain two
compare_group}

\item{min_gene_expressed}{Genes expressed in minimum number of cells}

\item{min_valid_cells}{Minimum number of genes detected in the cell}

\item{contrast}{String vector specifying the contrast to be
tested against the log2-fold-change threshold}

\item{verbose}{Whether show the progress of computing}

\item{data_type}{Type of data. Available options are:
\itemize{
 \item{'raw data'}: Raw count data without any pre-processing
 \item{'log count'}: Normalized and log-transformed data
}}
}
\value{
A matrix of differentially expressed genes and related statistics.
}
\description{
Identifies differentially expressed genes between two groups of cells using
a Wilcoxon Rank Sum test
}
